+++ src/common/DirectionalLight.hpp
@@ -11,7 +11,7 @@ public:
 	DirectionalLight(GLfloat red, GLfloat green, GLfloat blue,
 		GLfloat aIntensity, GLfloat dIntensity,
 		GLfloat xDir, GLfloat yDir, GLfloat zDir,
-		GLfloat shadowWidth, GLfloat shadowHeight);
+		GLuint shadowWidth, GLuint shadowHeight);
 
 	virtual void UseLight(GLuint ambientIntensityLocation, GLuint ambientColourLocation,
 		GLuint diffuseIntensityLocation, GLuint directionLocation);
