+++ src/common/Light.cpp
@@ -8,7 +8,7 @@ Light::Light()
 }
 
 Light::Light(GLfloat red, GLfloat green, GLfloat blue, GLfloat ambientIntensity, GLfloat diffuseIntensity,
-	GLfloat shadowWidth, GLfloat shadowHeight)
+	GLuint shadowWidth, GLuint shadowHeight)
 {
 	color = glm::vec3(glm::clamp(red, 0.0f, 1.0f),
 		glm::clamp(green, 0.0f, 1.0f),
